Split coordinator into a seperate file
