+++ homeassistant/components/homewizard_energy/__init__.py
@@ -1,6 +1,5 @@
 """The Homewizard Energy integration."""
 import asyncio
-from datetime import timedelta
 import logging
 
 import aiohwenergy
@@ -8,25 +7,11 @@ import async_timeout
 import voluptuous as vol
 
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_API_VERSION, CONF_ID, CONF_NAME, CONF_STATE
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
-from .const import (
-    ATTR_BRIGHTNESS,
-    ATTR_POWER_ON,
-    ATTR_SWITCHLOCK,
-    CONF_API,
-    CONF_DATA,
-    CONF_MODEL,
-    CONF_SW_VERSION,
-    CONF_UNLOAD_CB,
-    COORDINATOR,
-    DOMAIN,
-    MODEL_P1,
-    PLATFORMS,
-)
+from .const import CONF_API, CONF_UNLOAD_CB, COORDINATOR, DOMAIN, PLATFORMS
+from .coordinator import HWEnergyDeviceUpdateCoordinator as Coordinator
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -85,9 +70,9 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry):
             await energy_api.close()
 
     # Create coordinator
-    coordinator = hass.data[DOMAIN][entry.data["unique_id"]][
-        COORDINATOR
-    ] = HWEnergyDeviceUpdateCoordinator(hass, energy_api)
+    coordinator = hass.data[DOMAIN][entry.data["unique_id"]][COORDINATOR] = Coordinator(
+        hass, energy_api
+    )
     await coordinator.async_config_entry_first_refresh()
 
     hass.data[DOMAIN][entry.data["unique_id"]][CONF_API] = energy_api
@@ -123,73 +108,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry):
             unload_cb()
 
     return unload_ok
-
-
-class HWEnergyDeviceUpdateCoordinator(DataUpdateCoordinator):
-    """Gather data for the energy device."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        api: aiohwenergy.HomeWizardEnergy,
-    ) -> None:
-        """Initialize Update Coordinator."""
-
-        self.api = api
-
-        update_interval = self.get_update_interval()
-        super().__init__(hass, _LOGGER, name="", update_interval=update_interval)
-
-    def get_update_interval(self) -> timedelta:
-        """Return best interval for product type."""
-        try:
-            product_type = self.api.device.product_type
-        except AttributeError:
-            product_type = "Unknown"
-
-        if product_type == MODEL_P1:
-            try:
-                smr_version = self.api.data.smr_version
-                if smr_version == 50:
-                    return timedelta(seconds=1)
-
-            except AttributeError:
-                pass
-
-        return timedelta(seconds=5)
-
-    async def _async_update_data(self) -> dict:
-        """Fetch all device and sensor data from api."""
-        try:
-            async with async_timeout.timeout(10):
-                # Update all properties
-                status = await self.api.update()
-
-                if not status:
-                    raise Exception("Failed to fetch data")
-
-                data = {
-                    CONF_NAME: self.api.device.product_name,
-                    CONF_MODEL: self.api.device.product_type,
-                    CONF_ID: self.api.device.serial,
-                    CONF_SW_VERSION: self.api.device.firmware_version,
-                    CONF_API_VERSION: self.api.device.api_version,
-                    CONF_DATA: {},
-                    CONF_STATE: None,
-                }
-
-                for datapoint in self.api.data.available_datapoints:
-                    data[CONF_DATA][datapoint] = getattr(self.api.data, datapoint)
-
-                if self.api.state is not None:
-                    data[CONF_STATE] = {
-                        ATTR_POWER_ON: self.api.state.power_on,
-                        ATTR_SWITCHLOCK: self.api.state.switch_lock,
-                        ATTR_BRIGHTNESS: self.api.state.brightness,
-                    }
-
-        except Exception as ex:
-            raise UpdateFailed(ex) from ex
-
-        self.name = data[CONF_NAME]
-        return data
